Implemented Order entity class

+++ src/main/resources/db/migration/V6__create_order_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE customer_order(
-    order_id INTEGER GENERATED ALWAYS AS IDENTITY,
+    order_id BIGINT GENERATED ALWAYS AS IDENTITY,
     date_placed TIMESTAMP NOT NULL,
     order_total DOUBLE PRECISION NOT NULL,
     customer_id INTEGER NOT NULL,
